% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createGeneLists.R
\name{createGeneLists}
\alias{createGeneLists}
\title{Creates gene lists frrom Random Forest Model}
\usage{
createGeneLists(rfModel = rfmodel, numberPerList = 10)
}
\arguments{
\item{rfModel}{Random Forest Model}

\item{numberPerList}{Number of genes per classification to return. Default 10.}
}
\value{
A list of named gene lists
}
\description{
Creates gene lists frrom Random Forest Model
}
\author{
Kimberle Shen
}
