% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRFModel.R
\name{createRFModel}
\alias{createRFModel}
\title{Creates Random Forest model from Seurat Object's Idents}
\usage{
createRFModel(labelled.seurat.obj, nfeatures = 200)
}
\arguments{
\item{labelled.seurat.obj}{Seurat Object where Idents are the labels for training. Typically cell type/cell state.}

\item{nfeatures}{Number of most variable features to be used for training the Random Forest. Default 200.}
}
\value{
A list of two items, a Random Forest model and features used, to be used in predictCells()
}
\description{
Creates Random Forest model from Seurat Object's Idents
}
\author{
Kimberle Shen
}
