% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSeuratObjectPipeline.R
\name{createSeuratObjectPipeline}
\alias{createSeuratObjectPipeline}
\title{Creates Seurat object from 10X cell ranger output}
\usage{
createSeuratObjectPipeline(
  data.dir = "~/filtered_feature_bc_matrix",
  nFeature_RNA_lower = 500,
  nFeature_RNA_upper = 5000,
  percent.mt = 5,
  nfeatures = 2000,
  dims = 20,
  clusterResolution = 0.8
)
}
\arguments{
\item{data.dir}{Directory where data is stored. Default "~/filtered_feature_bc_matrix".}

\item{nFeature_RNA_lower}{Lower bound of nFeature_RNA. Default 500.}

\item{nFeature_RNA_upper}{Upper bound of nFeature_RNA. Default 5000.}

\item{percent.mt}{Upper bound of percentage of mitochondria genes. Default 5.}

\item{nfeatures}{Number of top most variable genes. Default 2000.}

\item{dims}{Number of dimensions for finding neighbors and UMAP. Default 20.}

\item{clusterResolution}{Cluster resolution for finding clusters. Default 0.8.}
}
\value{
Seurat Object with pca and UMAP calculated, ready for use with predictCells().
}
\description{
Creates Seurat object from 10X cell ranger output
}
\author{
Kimberle Shen
}
