% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleSeuratObjectLabelled.R
\docType{data}
\name{exampleSeuratObjectLabelled}
\alias{exampleSeuratObjectLabelled}
\title{Example Seurat Object - Labelled}
\format{
A Seurat Object with ~400 cells:
\describe{
  \item{Current Idents}{Cell Type (one of): Oligodendrocytes, Microglia, Astrocytes, Endothelial, Neurons, OPCs, Tcells, VSMCs}
  \item{Dimensional Reductions}{umap and pca calculated}
  \item{contains}{contains 8 cell types of ~50 cells each}
}
}
\usage{
exampleSeuratObjectLabelled
}
\description{
A dataset containing ~400 cells.
}
\keyword{datasets}
