% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleSeuratObjectUnlabelled.R
\docType{data}
\name{exampleSeuratObjectUnlabelled}
\alias{exampleSeuratObjectUnlabelled}
\title{Example Seurat Object - Unlabelled}
\format{
A Seurat Object with ~400 cells:
\describe{
  \item{Current Idents}{UMAP clusters 0-9}
  \item{Dimensional Reductions}{umap and pca calculated}
}
}
\usage{
exampleSeuratObjectUnlabelled
}
\description{
A dataset containing ~400 cells.
}
\keyword{datasets}
