% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictCells.R
\name{predictCells}
\alias{predictCells}
\title{Takes a Seurat Object and returns the Seurat Object with predicted CNS cell types as Idents}
\usage{
predictCells(seuratObj, rfModel = rfmodel)
}
\arguments{
\item{seuratObj}{Seurat Object.}

\item{rfModel}{Random Forest Object from randomForest package. Default: A Random Forest Object trained with CNS cell types.}
}
\value{
Seurat Object with predicted cell types as Idents and stored under "predictedCellType" metadata
}
\description{
Takes a Seurat Object and returns the Seurat Object with predicted CNS cell types as Idents
}
\author{
Kimberle Shen
}
