% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellMarkerPlots.R
\name{cellMarkerPlots}
\alias{cellMarkerPlots}
\title{Creates Feature Plots based on average gene expression of a list of genes}
\usage{
cellMarkerPlots(seuratObj, geneList = gsl)
}
\arguments{
\item{seuratObj}{Seurat Object}

\item{geneList}{A list of lists containing gene names}
}
\value{
Feature plot with gene set scores
}
\description{
Creates Feature Plots based on average gene expression of a list of genes
}
\author{
Kimberle Shen
}
