% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictCells.R
\name{predictCells}
\alias{predictCells}
\title{Takes a Seurat Object and returns the Seurat Object with predicted cell types as Idents}
\usage{
predictCells(seuratObj, rfModel = mouseBrain)
}
\arguments{
\item{seuratObj}{Seurat Object.}

\item{rfModel}{Random Forest Object from randomForest package. In-built options include mouseBrain (default), mouseLiver, mouseKidney, mouseHeart, mouseMuscle, humanPBMC. You can also use your own Random Forest object created with createRFModel().}
}
\value{
Seurat Object with predicted cell types as Idents and stored under "predictedCellType" metadata
}
\description{
Takes a Seurat Object and returns the Seurat Object with predicted cell types as Idents
}
\author{
Kimberle Shen
}
